requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
appdirs==1.4.4
asyncio-dgram==2.1.2
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.11.17
charset-normalizer==3.3.2
cssselect==1.2.0
discord==2.3.2
discord.py==2.3.2
dnspython==2.4.2
fake-useragent==1.4.0
frozenlist==1.4.1
idna==3.6
importlib-metadata==7.0.1
lxml==4.9.4
mcstatus==11.1.0
multidict==6.0.4
parse==1.20.0
pyee==8.2.2
pyppeteer==1.0.2
pyquery==2.0.0
python-dotenv==1.0.0
requests==2.31.0
requests-html==0.10.0
soupsieve==2.5
tqdm==4.66.1
urllib3==1.26.18
w3lib==2.1.2
websockets==10.4
yarl==1.9.4
zipp==3.17.0
